OMS7-N: Disable Zygote preloaded drawables [10/14]
With a theme applied, most of these preloaded drawables go unused.  Any
assets the theme has redirected will need to be loaded with each app
process regardless.  Worse, preloads make it impossible to do asset
redirection for constituent parts of a preloaded drawable (for instance,
individual states of a StateListDrawable cannot be redirected).

Some day it might be nice to revisit this and see if there's a way to
reintroduce the drawable cache in a way that can be altered at runtime
without significant complexity or runtime penalty.

Change-Id: I253b1a22482ac664c196533a4c2fcd88ae84b996